Refactored to avoid circular dependency
